+++ ios/RNSentry.mm
@@ -7,6 +7,14 @@
 #import "RCTConvert.h"
 #endif
 
+#if __has_include(<hermes/hermes.h>)
+#define SENTRY_PROFILING_ENABLED 1
+#import <Sentry/SentryProfilingConditionals.h>
+#else
+#define SENTRY_PROFILING_ENABLED 0
+#define SENTRY_TARGET_PROFILING_SUPPORTED 0
+#endif
+
 #import <Sentry/Sentry.h>
 #import <Sentry/PrivateSentrySDKOnly.h>
 #import <Sentry/SentryScreenFrames.h>
@@ -14,12 +22,7 @@
 #import <Sentry/SentryBinaryImageCache.h>
 #import <Sentry/SentryDependencyContainer.h>
 #import <Sentry/SentryFormatter.h>
-
-#if __has_include(<hermes/hermes.h>)
-#define SENTRY_PROFILING_ENABLED 1
-#else
-#define SENTRY_PROFILING_ENABLED 0
-#endif
+#import <Sentry/SentryCurrentDateProvider.h>
 
 // This guard prevents importing Hermes in JSC apps
 #if SENTRY_PROFILING_ENABLED
@@ -594,16 +597,40 @@ RCT_EXPORT_METHOD(enableNativeFramesTracking)
 }
 
 static NSString* const enabledProfilingMessage = @"Enable Hermes to use Sentry Profiling.";
+static SentryId* nativeProfileTraceId = nil;
+static uint64_t nativeProfileStartTime = 0;
 
 RCT_EXPORT_SYNCHRONOUS_TYPED_METHOD(NSDictionary *, startProfiling)
 {
 #if SENTRY_PROFILING_ENABLED
     try {
         facebook::hermes::HermesRuntime::enableSamplingProfiler();
+        if (nativeProfileTraceId == nil && nativeProfileStartTime == 0) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+            nativeProfileTraceId = [[SentryId alloc] init];
+            nativeProfileStartTime = [PrivateSentrySDKOnly startProfilerForTrace: nativeProfileTraceId];
+#endif
+        } else {
+            NSLog(@"Native profiling already in progress. Currently existing trace: %@", nativeProfileTraceId);
+        }
         return @{ @"started": @YES };
     } catch (const std::exception& ex) {
+        if (nativeProfileTraceId != nil) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+            [PrivateSentrySDKOnly discardProfilerForTrace: nativeProfileTraceId];
+#endif
+            nativeProfileTraceId = nil;
+        }
+        nativeProfileStartTime = 0;
         return @{ @"error": [NSString stringWithCString: ex.what() encoding:[NSString defaultCStringEncoding]] };
     } catch (...) {
+        if (nativeProfileTraceId != nil) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+            [PrivateSentrySDKOnly discardProfilerForTrace: nativeProfileTraceId];
+#endif
+            nativeProfileTraceId = nil;
+        }
+        nativeProfileStartTime = 0;
         return @{ @"error": @"Failed to start profiling" };
     }
 #else
@@ -615,6 +642,17 @@ RCT_EXPORT_SYNCHRONOUS_TYPED_METHOD(NSDictionary *, stopProfiling)
 {
 #if SENTRY_PROFILING_ENABLED
     try {
+        NSDictionary<NSString *, id> * nativeProfile = nil;
+        if (nativeProfileTraceId != nil && nativeProfileStartTime != 0) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+            uint64_t nativeProfileStopTime = [[[SentryDependencyContainer sharedInstance] dateProvider] systemTime];
+            nativeProfile = [PrivateSentrySDKOnly collectProfileBetween:nativeProfileStartTime and:nativeProfileStopTime forTrace:nativeProfileTraceId];
+#endif
+        }
+        // Cleanup native profiles
+        nativeProfileTraceId = nil;
+        nativeProfileStartTime = 0;
+
         facebook::hermes::HermesRuntime::disableSamplingProfiler();
         std::stringstream ss;
         facebook::hermes::HermesRuntime::dumpSampledTraceToStream(ss);
@@ -633,10 +671,35 @@ RCT_EXPORT_SYNCHRONOUS_TYPED_METHOD(NSDictionary *, stopProfiling)
         }
 #endif
 
-        return @{ @"profile": data };
+        if (data == nil) {
+          return @{ @"error": @"Failed to retrieve Hermes profile." };
+        }
+
+        if (nativeProfile == nil) {
+          return @{ @"profile": data };
+        }
+
+        return @{
+          @"profile": data,
+          @"nativeProfile": nativeProfile,
+        };
     } catch (const std::exception& ex) {
+        if (nativeProfileTraceId != nil) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+          [PrivateSentrySDKOnly discardProfilerForTrace: nativeProfileTraceId];
+#endif
+          nativeProfileTraceId = nil;
+        }
+        nativeProfileStartTime = 0;
         return @{ @"error": [NSString stringWithCString: ex.what() encoding:[NSString defaultCStringEncoding]] };
     } catch (...) {
+        if (nativeProfileTraceId != nil) {
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+          [PrivateSentrySDKOnly discardProfilerForTrace: nativeProfileTraceId];
+#endif
+          nativeProfileTraceId = nil;
+        }
+        nativeProfileStartTime = 0;
         return @{ @"error": @"Failed to stop profiling" };
     }
 #else